+++ frontend/src/pages/Home.js
@@ -3,9 +3,7 @@ import WelcomeMessage from "./WelcomeMessage";
 
 function Home() {
   return (
-    <div
-      className="flex flex-col h-screen bg-cover bg-center bg-no-repeat font-nosifer"
-    >
+    <div className="flex flex-col h-screen bg-cover bg-center bg-no-repeat font-nosifer">
       <WelcomeMessage />
     </div>
   );
